Fix some type conversion issues brought up by static analysis.

+++ src/libdecaf/src/gpu/opengl/opengl_constants.h
@@ -6,7 +6,7 @@ namespace gpu
 namespace opengl
 {
 
-extern size_t
+extern unsigned
 MaxUniformBlockSize;
 
 } // namespace opengl
